extract firebase init
