+++ App/Containers/RootContainer.js
@@ -2,20 +2,23 @@
 
 import React, { Component } from 'react';
 import { View, StatusBar } from 'react-native';
-import NavigationRouter from '../Navigation/NavigationRouter';
 import { connect } from 'react-redux';
+
+import NavigationRouter from '../Navigation/NavigationRouter';
 import StartupActions from '../Redux/StartupRedux';
 import ReduxPersist from '../Config/ReduxPersist';
 
-// Styles
 import styles from './Styles/RootContainerStyle';
 
 class RootContainer extends Component {
   componentDidMount() {
-    // if redux persist is not active fire startup action
     if (!ReduxPersist.active) {
       this.props.startup();
     }
+  }
+
+  props: {
+    startup: () => void;
   }
 
   render() {
@@ -28,7 +31,6 @@ class RootContainer extends Component {
   }
 }
 
-// wraps dispatch to create nicer functions to call within our component
 const mapDispatchToProps = dispatch => ({
   startup: () => dispatch(StartupActions.startup()),
 });
