+++ App/Services/ImmutablePersistenceTransform.js
@@ -28,7 +28,7 @@ export default {
     // Another equal terrifying option would be to try to pass their other check
     // which is lodash isPlainObject.
     // --- END HACKZORZ ---
-    state.mergeDeep = R.identity;
+    state.mergeDeep = R.identity; // eslint-disable-line
     return toImmutable(state);
   },
   in: (raw: Object) =>
