+++ App/Services/RehydrationServices.js
@@ -1,6 +1,7 @@
-import ReduxPersist from '../Config/ReduxPersist';
 import { AsyncStorage } from 'react-native';
 import { persistStore } from 'redux-persist';
+
+import ReduxPersist from '../Config/ReduxPersist';
 import StartupActions from '../Redux/StartupRedux';
 
 const updateReducers = (store: Object) => {
@@ -20,7 +21,6 @@ const updateReducers = (store: Object) => {
         preview: 'Reducer Version Change Detected',
         important: true,
       });
-      // Purge store
       persistStore(store, config, startup).purge();
       AsyncStorage.setItem('reducerVersion', reducerVersion);
     } else {
