+++ App/Services/batches/addChannels.js
@@ -1,6 +1,6 @@
 // @flow
 import Promise from 'bluebird';
-import * as firebase from 'firebase';
+import {firebaseApp} from '../firebase/';
 
 import {ChannelsResource, channelsRef, logFinished, snapshotExists} from '../index';
 import type {TChannel} from '../../types/Channel';
@@ -9,13 +9,13 @@ import type {TChannelResponse} from '../../types/ChannelResponse';
 // add CHANNEL_IDS before dispatching batch
 const CHANNEL_IDS = `
   UC2rbyOa3Jo7vGSibqKcRjqw, UCI5qMix97T3tVZfxmHObDjA, UC0elp2101KAxbaAMzInGerA, UCB10yM5qyQpNsMNDQ1VHxKg, UCiOm_FmFK4jxB9VRuFC1pag,
-  UCCuizDTLsr-mNm_PEGdChVg, UCRdQOBEzSpAOtZ_yitD7GpA, UCFTVNLC7ysej-sD5lkLqNGA, UC5xAkS4828lDivq8cKFGSyw, UC__AsSnEuyVgO9TWvZE_ziA,
+  UCCuizDTLsr-mNm_PEGdChVg, UCRdQOBEzSpAOtZ_yitD7GpA, UCFTVNLC7ysej-sD5lkLqNGA, UC__AsSnEuyVgO9TWvZE_ziA, UCOZ7Kq5_VWBC-TtteAcsRBg,
   UCfCY70zRsvnnKzQ39mBq0rw, UCr-QcqNToYablI-jU2VPVSw, UCQFvuGmxBmv-c_irVda7mwg, UCLfbmGhvu7xTHXktmMWdfvw, UC69xoWl5-Y3m-oRFTORIbKw,
   UCAjZx0WhDOjIWsoy0owuK4w, UC0-H-XRuSPbDdPRmAnuzxSQ, UC97ysepeThCEbhrWhYrqxAw, UCjeKoCr7YTNn0bmMO1HdyVw, UCrd9neCNtPkvQukpNTYrP4Q,
   UCLi8qtd3QLvXCZLyobiodiw, UCtVnAr55ALHAuI7Yyz4GZWg, UCC1BNMUl5dnju1b9oKpkysg, UC5vDiFXIokkQv5xZ91fg-oA, UCgVA07MrT-XCN1t8WMeoFcA
   UCsTM1roCxoot1-03EO5zQxg, UCIyMwXronD5pT5cx-G_KSHA, UCmsA3A5_HKBwI9OktSttTFg, UC8X6Cb-pumA1CNkd3K-lCyA, UCZvbol2FjFRiqc0hwIvpmMQ,
   UCN559lrbV9wt46NwlnPJtPw, UCNIwy_Q7EjUxLlsewfuhjgg, UC7fN-mbfjZVUo755gu-1VuQ, UChhn3SKUYAcajcG9mn7dGqA, UCHPARXha7xQNbfzwzY7NzUw,
-  UClW0iupPReSXYDmlvwYfQhg, UC7rqz5As19qYWl2Rc4z-iig, UCMyArXimoNAZjE1P2xLES7A, UCNS93jTKjnzNG7pYUCN5PrQ
+  UClW0iupPReSXYDmlvwYfQhg, UC7rqz5As19qYWl2Rc4z-iig, UCMyArXimoNAZjE1P2xLES7A, UCNS93jTKjnzNG7pYUCN5PrQ, UCDRWhpjy_lU_phS3O31AjRQ,
 `;
 
 const createChannel = (channel: TChannelResponse): TChannel => {
@@ -27,8 +27,8 @@ const createChannel = (channel: TChannelResponse): TChannel => {
 
   return {
     id: '',
-    createdAt: firebase.database.ServerValue.TIMESTAMP,
-    modifiedAt: firebase.database.ServerValue.TIMESTAMP,
+    createdAt: firebaseApp.database.ServerValue.TIMESTAMP,
+    modifiedAt: firebaseApp.database.ServerValue.TIMESTAMP,
     rank: 0,
     score: subscriberCount,
     likeCount: 0,
