+++ App/Services/firebase/index.js
@@ -1,3 +1,4 @@
 // @flow
+export * from './init';
 export * from './ref';
 export * from './subscribe';
