+++ App/Services/firebase/ref.js
@@ -1,17 +1,6 @@
 // @flow
-import * as firebase from 'firebase';
-import Secrets from 'react-native-config';
 import type {TChannel} from '../../types/Channel';
-
-const firebaseConfig = {
-  apiKey: Secrets.FIREBASE_API_KEY,
-  authDomain: Secrets.FIREBASE_AUTH_DOMAIN,
-  databaseURL: Secrets.FIREBASE_DATABASE_URL,
-  storageBucket: Secrets.FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: Secrets.FIREBASE_MESSAGING_SENDER_ID,
-};
-
-firebase.initializeApp(firebaseConfig);
+import {firebaseApp} from './init';
 
 export type TChannelsRef = {
   startAt: () => TChannelsRef;
@@ -24,5 +13,5 @@ export type TChannelsRef = {
   on: () => void;
 }
 
-export const channelsRef: TChannelsRef = firebase.database().ref('channels');
-export const likesRef = firebase.database().ref('likes');
+export const channelsRef: TChannelsRef = firebaseApp.database().ref('channels');
+export const likesRef = firebaseApp.database().ref('likes');
