+++ doc/model.md
@@ -1,15 +1,18 @@
 
-* `/resource/firebase_id/`は、便宜上配列として表している
+* `/resource/firebaseId/`は、便宜上配列として表している
 
 ```JavaScript
 /*
 like登録
   channels.like_countを+1
   likes.user_id.countを+1
+
   user.possible_like_countを-1。0になったらlike不可に。
+
 like削除
   channels.like_countを-1
   likes.user_id.countを-1
+
   user.possible_like_countを+1。20になったらそれ以上は回復しない。
 */
 ```
@@ -17,7 +20,7 @@ like削除
 ```JavaScript
 /*
 マイランキングページ
-ref.child(likes/user_id).get(data => {
+ref.child('/likes/user_id/').once(data => {
   ref.child(channels/data.channel_id).get();
 });
 */
@@ -25,7 +28,7 @@ export type likes = {
   user_id: [
     {
       id: string;
-      channel_id: string;
+      channelId: string;
       rank: number;
       likeCount: number;
     }
